+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Release Notes
 
 ## [Unreleased](https://github.com/laravel/sanctum/compare/v2.10.0...2.x)
-
+### Added
+- `Sanctum::$validateCallback` callback for more granular control over access token validation
 
 ## [v2.10.0 (2021-04-20)](https://github.com/laravel/sanctum/compare/v2.9.4...v2.10.0)